requirements.txt
aiohttp==3.8.4
aioredis==2.0.1
aiosignal==1.3.1
alembic==1.12.1
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
bcrypt==4.0.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloud-sql-python-connector==1.2.1
cloudinary==1.37.0
cryptography==40.0.1
databases==0.7.0
distro==1.9.0
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
exceptiongroup==1.1.1
fastapi==0.94.0
frozenlist==1.3.3
google-auth==2.17.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.2
multidict==6.0.4
openai==1.6.1
orjson==3.8.7
packaging==23.0
passlib==1.7.4
pluggy==1.0.0
psycopg2-binary==2.9.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.6
pytest==7.2.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
redis==3.5.3
rejson==0.5.6
requests==2.28.2
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.26.0.post1
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.9.0
ujson==5.7.0
urllib3==1.26.15
uvicorn==0.21.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
yarl==1.8.2
